requirements.txt
Django==5.1.1
djangorestframework==3.15.2
elasticsearch==8.15.1
redis==5.0.8
